requirements.txt
Django==2.2.16
djangorestframework==3.12.4
requests==2.26.0
python-dotenv
psycopg2